Make git prompt faster with submodules.
This change makes it so we only check for changes in the commit HEAD for
submodules, rather than looking for actual changes. This should be fine
because I don't really work with submodules as part of normal
development, but only a way to include dependencies in git repos.
Therefore I only really care about when those deps are updated.

A slightly better (but slower) change would be to use
`--ignore_submodules=dirty`, which checks for changes in tracked files
but does not have to scan the whole subtree for untracked files.