Fix kakoune config for missing plugins.
Apparently the maintainer of connect.kak and plug.kak completely deleted
their repos >:(.

For plug.kak, the alternative is probably to  migrate back to andreyorst's
plug.kak, which is maintained again. However, I don't want to deal with
that so for now I am forking the old plug.kak directly into my autoload.

For connect.kak, the "correct" thing to do is migrate to kakoune.cr, but
that requires installing crystal, which I have zero interest in.
Fortunately, a fork of connect.kak is available on github already.